pyproject.toml
[project]
name = "bielik-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.16.0",
    "python-dotenv>=1.0.1",
    "replicate>=1.0.4",
]
